Add types for cardTask
